#31: Fix minor templating problems in jules docs
This fixes a problem which caused the table of contents to disappear
from the sidebar at vn5.5.  It also updates the copyright year and
moves the jules-tweaks.css directive into conf.py.

These changes allow layout.html to be retired.  This removes the
reference to Crown Copyright in the new docs which should be more in
line with the BSD-3 licence.