Add a deploy step to the workflow
This adds a script which carries out the actions required to build and
release a new version of the jules documentation when the action is a
push to the master branch if they do not exist for the most recent
release tag.

This partitions the workflow to trigger the build-only step when the
pushing to a branch or on a PR and only the deployment step when
pushing to master.