Deploy documentation updates to gh-pages
Change the workflow and the release.py script to deploy to a gh-pages
branch rather than using the master branch as is presently the case.
This should bypass problems associated with the branch controls on
main.

In order to take advantage of this change, the github repo settings
need to be updated to serve the documentation from the gh-pages
branch.