Use a persistent gh-pages branch for documentation
Prompted by comments from a reviewer, this modifies the workflow to
use a persistent gh-pages branch to hold the documentation.  This
ensures that links created when a version tag is created endure across
multiple releases.  This is achieved by attempting to check out an
existing gh-pages branch and only using checkout -b when this fails.

This also adds the --force flag to the push command that updates the
upstream branch to deal with spurious errors which result from the
merger of the trunk into the docs branch.