Update the conda environment to newer versions
Upgrade to the very latest version of sphinx because this adds the
ability to ignore a URL after it has been redirected.  This feature is
required to deal with spurious failures that occur when running from
github because journal URLs cannot be resolved without a subscription.

This exact problem was raised in this issue:

https://github.com/sphinx-doc/sphinx/issues/11233

And resolved by this PR:

https://github.com/sphinx-doc/sphinx/pull/13127